Refactor LogHelper
